main to side transition update

+++ all_create_ff.py
@@ -48,10 +48,9 @@ def AllVideo(main_end,pickArr,usedTime):
     if not os.path.exists(filePath):
         os.mkdir(filePath)
     AllArr.append(ffVideo("./template/pass/passMain.mp4"))
-    for clips in range(main_end,1,-1):
-        AllArr.append(ffVideo(f"./output_clips/MainRank_{clips}.mp4"))
-        AllArr.append(ffVideo("./template/pass/pass.mp4"))
-    AllArr.append(ffVideo("./output_clips/MainRank_1.mp4"))
+
+    # Pick Up
+
     if pickArr != []:
         AllArr.append(ffVideo("./template/pass/passPick.mp4"))
         for clipsto in range(1,len(pickArr)):
@@ -65,12 +64,23 @@ def AllVideo(main_end,pickArr,usedTime):
             pass
         AllArr.append(inoutVideo("./template/pass/passCanbin.mp4"))
         AllArr.append(inoutVideo("./custom/canbin.mp4"))
-        AllArr.append(inVideo("./template/pass/passSide.mp4"))
+        AllArr.append(inVideo("./template/pass/passMain.mp4"))
     else:
         if pickArr != []:
             AllArr.append(ffVideo(f"./output_clips/PickRank_{len(pickArr)}.mp4"))
+        AllArr.append(ffVideo("./template/pass/passMain.mp4"))
+
+    # 倒数五位主榜
+
+    for clips in range(5,1,-1):
+        AllArr.append(ffVideo(f"./output_clips/MainRank_{clips}.mp4"))
+        AllArr.append(ffVideo("./template/pass/pass.mp4"))
+    AllArr.append(ffVideo("./output_clips/MainRank_1.mp4"))
+
+    if os.path.exists("./custom/ed.mp4"):
         AllArr.append(ffVideo("./template/pass/passSide.mp4"))
-    AllArr.append(ffVideo("./output_clips/SideRank.mp4"))
+        AllArr.append(ffVideo("./output_clips/SideRank.mp4"))
+
     onum = 0
     for items in AllArr:
         onum += 1
@@ -78,7 +88,7 @@ def AllVideo(main_end,pickArr,usedTime):
             combVideo = items
             continue
         combVideo = ffmpeg.concat(combVideo[0],combVideo[1],items[0],items[1],v=1,a=1).node
-    ffmpeg.output(combVideo[0],combVideo[1],f'./output_all/Rank_{usedTime}.mp4',**render_format).run()
+    ffmpeg.output(combVideo[0],combVideo[1],f'./output_all/Rank_{usedTime}.mp4',**all_render_format).run()
     if os.path.exists("./custom/canbin.mp4"):
         shutil.move("./custom/canbin.mp4",f"{filePath}/canbin_{usedTime}.mp4")
     shutil.move("./output_clips/Opening.mp4",f"{filePath}/Opening.mp4")
@@ -86,6 +96,7 @@ def AllVideo(main_end,pickArr,usedTime):
         shutil.move(f"./output_clips/MainRank_{clips}.mp4",f"{filePath}/MainRank_{clips}.mp4")
     for clipsto in range(1,len(pickArr)+1):
         shutil.move(f"./output_clips/PickRank_{clipsto}.mp4",f"{filePath}/PickRank_{clipsto}.mp4")
-    shutil.move("./output_clips/SideRank.mp4",f"{filePath}/SideRank.mp4")
-    os.remove("./custom/ed.mp4")
+    if os.path.exists("./custom/ed.mp4"):
+        shutil.move("./output_clips/SideRank.mp4",f"{filePath}/SideRank.mp4")
+        os.remove("./custom/ed.mp4")
     os.remove("./custom/time.txt")
